+++ Ristorante/src/main/java/model/Pagamento.java
@@ -3,11 +3,11 @@ package model;
 public class Pagamento {
 	private int id_pagamento;
 	private int id_tavolo;
-	private double costo_totale;
+	private float costo_totale;
 	private String stato;
 	private String data;
 
-	public Pagamento(int id,int id_tavolo, double ct, String s, String d) {
+	public Pagamento(int id,int id_tavolo, float ct, String s, String d) {
 		this.id_pagamento=id;
 		this.id_tavolo = id_tavolo;
 		this.costo_totale = ct;
@@ -34,11 +34,11 @@ public class Pagamento {
 		this.id_tavolo = id_tavolo;
 	}
 
-	public double getCosto_totale() {
+	public float getCosto_totale() {
 		return costo_totale;
 	}
 
-	public void setCosto_totale(double costo_totale) {
+	public void setCosto_totale(float costo_totale) {
 		this.costo_totale = costo_totale;
 	}
